% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_global_options}
\alias{get_global_options}
\title{Read in permanent options}
\usage{
get_global_options(
  options = c("nab", "outcomes", "learners", "cvtune", "cvperf", "nfolds",
    "combination_method", "binary_outcomes", "importance_grp", "importance_ind",
    "ind_importance_type", "report_name", "return", "sens_thresh", "multsens_nab",
    "same_subset", "var_thresh", "output", "slfits", "analysis"),
  options_boolean = c(FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE,
    FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE)
)
}
\arguments{
\item{options}{A character vector.}

\item{options_boolean}{A booloean vector.}
}
\value{
Named list of system variables
}
\description{
Read in permanent options
}
\examples{
get_global_options(c("nab", "cvtune"), c(FALSE, TRUE))
}
