% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_intrinsic_importance_table_description}
\alias{get_intrinsic_importance_table_description}
\title{get intrinsic importance table text}
\usage{
get_intrinsic_importance_table_description(
  opts,
  any_cont = TRUE,
  any_dich = TRUE,
  cont_nms = c("IC$_{50}$", "IC$_{80}$", "IIP"),
  bin_nms = c("Sensitivity"),
  num_obs_fulls = 100,
  num_obs_reds = 100,
  n_row_now = 100,
  vimp_threshold = 0.05,
  importance_type = "marg",
  any_signif = FALSE
)
}
\arguments{
\item{opts}{options}

\item{cont_nms}{nice names of continuous outcomes}

\item{bin_nms}{nice names of binary outcomes}

\item{num_obs_fulls}{number of total obs for "full" regression (may differ for each outcome)}

\item{num_obs_reds}{number of total obs for reduced regression (may differ for each outcome)}

\item{n_row_now}{(may differ for each outcome)}

\item{importance_type}{"marginal" or "conditional"}
}
\description{
get intrinsic importance table text
}
