% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_sys_var}
\alias{get_sys_var}
\title{Get system variables}
\usage{
get_sys_var(option = "nab", boolean = FALSE)
}
\arguments{
\item{option}{A character string.}

\item{boolean}{A booloean.}
}
\value{
Required parameters as expected by program
}
\description{
Get system variables
}
\examples{
get_sys_var("nab", FALSE)
get_sys_var("cvtune", TRUE)
}
