% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{make_nice_outcome}
\alias{make_nice_outcome}
\title{Printing nice outcomes,}
\usage{
make_nice_outcome(outcome = "ic50")
}
\arguments{
\item{options}{A character vector.}

\item{options_boolean}{A booloean vector.}
}
\value{
Named list of system variables
}
\description{
Printing nice outcomes,
}
\examples{
get_global_options(c("nab", "cvtune"), c(FALSE, TRUE))
}
