% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_super_learner_libraries.R
\name{make_screen_wrapper}
\alias{make_screen_wrapper}
\title{read a screen wrapper into global environment}
\usage{
make_screen_wrapper(var_thresh)
}
\description{
read a screen wrapper into global environment
}
