% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_super_learner_libraries.R
\name{sl_one_outcome}
\alias{sl_one_outcome}
\title{function to run super learner and cv super learner on a single outcome}
\usage{
sl_one_outcome(
  complete_dat,
  outcome_name,
  pred_names,
  opts,
  h2o_here,
  save_dir = Sys.getenv("slfits"),
  fit_name = paste0("fit_", outcome_name, ".rds"),
  cv_fit_name = paste0("cvfit_", outcome_name, ".rds"),
  save_full_object = TRUE,
  full_fit = TRUE,
  ss_folds = rep(1, opts$nfolds),
  SL.library = "SL.mean",
  call_out = FALSE,
  ...
)
}
\arguments{
\item{complete_dat}{the full dataset}

\item{outcome_name}{String name of outcome}

\item{pred_names}{Vector of string names of predictor variables}

\item{opts}{List of options outputted by get_global_options()}

\item{save_dir}{name of directory to save results to}

\item{fit_name}{name of fits (defaults to fit_<outcome_name>.rds)}

\item{cv_fit_name}{name of CV fits (defaults to cvfit_<outcome_name>.rds)}

\item{save_full_object}{Flag for whether or not to save the full fitted object, or just the fitted values}

\item{full_fit}{is it the full fit (TRUE) or a reduced fit (FALSE)?}

\item{SL.library}{the library to pass to SuperLearner}

\item{...}{additional arguments to pass to individual algorithms or the SuperLearner}
}
\description{
function to run super learner and cv super learner on a single outcome
}
